requirements.txt
discord.py[voice]==2.4.0
python-dotenv==1.0.1
jishaku==2.5.2
